Move app_entrypoint back to server
